Fixed the infinite rerender of code
