+++ package.json
@@ -54,6 +54,7 @@
     "framer-motion": "^11.3.31",
     "highlight.js": "^11.10.0",
     "jest": "^29.6.2",
+    "lodash": "^4.17.21",
     "lucide-react": "^0.471.1",
     "next": "15.1.4",
     "next-mdx-remote": "^5.0.0",
@@ -77,6 +78,7 @@
   },
   "devDependencies": {
     "@tailwindcss/typography": "^0.5.9",
+    "@types/lodash": "^4.17.14",
     "prettier": "3.3.3"
   },
   "packageManager": "yarn@1.22.22",
